Changed wording of first question response
Trying to duplicate as little of this logic as possible since this is
the stuff that's always changing and I'm likely to break something in
the process.

Needed to add "According to the state" to the first question responses
because "we" don't think that even fixes all of the problems.